# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mithra AI Family Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUserView"
    post:
      operationId: adminCreateUser
      tags: [admin]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserView"
  /admin/users/{userId}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserView"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/ai-keys:
    get:
      operationId: adminGetAiKeys
      tags: [admin]
      summary: Get AI provider API keys (admin only)
      responses:
        "200":
          description: AI keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiKeys"
    patch:
      operationId: adminUpdateAiKeys
      tags: [admin]
      summary: Update AI provider API keys (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiKeysUpdate"
      responses:
        "200":
          description: Updated keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiKeys"
...
  /admin/audit-logs:
    get:
      operationId: adminListAuditLogs
      tags: [admin]
      summary: List audit logs (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
  /admin/default-ai-config:
    get:
      operationId: adminGetDefaultAiConfig
      tags: [admin]
      summary: Get default AI config (admin only)
      responses:
        "200":
          description: Default AI config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConfig"
    patch:
      operationId: adminUpdateDefaultAiConfig
      tags: [admin]
      summary: Update default AI config for all users (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiConfigUpdate"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConfig"
components:
  schemas:
...
    AiConfig:
      type: object
      required: [id, defaultModel, temperature, maxTokens, streamingEnabled]
      properties:
        id:
          type: integer
        defaultModel:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
        streamingEnabled:
          type: boolean
        systemPrompt:
          type: ["string", "null"]
    AiConfigUpdate:
      type: object
      properties:
        defaultModel:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
        streamingEnabled:
          type: boolean
        systemPrompt:
          type: ["string", "null"]
...
    AiKeys:
      type: object
      properties:
        openaiKeySet:
          type: boolean
        anthropicKeySet:
          type: boolean
        geminiKeySet:
          type: boolean
        groqKeySet:
          type: boolean
        openrouterKeySet:
          type: boolean
    AiKeysUpdate:
      type: object
      properties:
        openaiKey:
          type: ["string", "null"]
        anthropicKey:
          type: ["string", "null"]
        geminiKey:
          type: ["string", "null"]
        groqKey:
          type: ["string", "null"]
        openrouterKey:
          type: ["string", "null"]

    # ─── Dashboard ────────────────────────────────────────────────────────────
...
    AdminUserView:
      type: object
      required: [id, clerkId, email, name, role, isActive, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        isActive:
          type: boolean
        totalChats:
          type: integer
        totalMessages:
          type: integer
        totalFiles:
          type: integer
        tokensUsed:
          type: integer
        createdAt:
          type: string
        lastActive:
          type: ["string", "null"]
    AdminUserInput:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
    AdminUserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
        isActive:
          type: boolean
...
    AuditLog:
      type: object
      required: [id, userId, action, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: ["string", "null"]
        action:
          type: string
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        details:
          type: ["string", "null"]
        createdAt:
          type: string