# lib/api-zod/src/generated/types/adminUserInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Mithra AI Family Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserInputRole } from './adminUserInputRole';

export interface AdminUserInput {
  email: string;
  name: string;
  role?: AdminUserInputRole;
}