# lib/api-zod/src/generated/types/aiConfig.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Mithra AI Family Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface AiConfig {
  id: number;
  defaultModel: string;
  temperature: number;
  maxTokens: number;
  streamingEnabled: boolean;
  /** @nullable */
  systemPrompt?: string | null;
}