# lib/api-zod/src/generated/types/aiKeysUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Mithra AI Family Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface AiKeysUpdate {
  /** @nullable */
  openaiKey?: string | null;
  /** @nullable */
  anthropicKey?: string | null;
  /** @nullable */
  geminiKey?: string | null;
  /** @nullable */
  groqKey?: string | null;
  /** @nullable */
  openrouterKey?: string | null;
}